Passing event up to stateful component from outfit card
